Create new filestructure
Move old PolarRobotics.cpp/h to Drive.cpp/h
Create new Robot and Drive cpps as discussed during meeting
